// src/api/get-tasks/get-tasks.generated.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string };
  String: { input: string; output: string };
  Boolean: { input: boolean; output: boolean };
  Int: { input: number; output: number };
  Float: { input: number; output: number };
  DateTime: { input: any; output: any };
};
// ...
export enum PointEstimate {
  Eight = "EIGHT",
  Four = "FOUR",
  One = "ONE",
  Two = "TWO",
  Zero = "ZERO",
}

export type Query = {
  __typename?: "Query";
  profile: User;
  tasks: Array<Task>;
  users: Array<User>;
};

export type QueryTasksArgs = {
  input: FilterTaskInput;
};

/** Status for Task */
export enum Status {
  Backlog = "BACKLOG",
  Cancelled = "CANCELLED",
  Done = "DONE",
  InProgress = "IN_PROGRESS",
  Todo = "TODO",
}

export type Task = {
  __typename?: "Task";
  assignee?: Maybe<User>;
  createdAt: Scalars["DateTime"]["output"];
  creator: User;
  dueDate: Scalars["DateTime"]["output"];
  id: Scalars["ID"]["output"];
  name: Scalars["String"]["output"];
  pointEstimate: PointEstimate;
  position: Scalars["Float"]["output"];
  status: Status;
  tags: Array<TaskTag>;
};

/** Enum for tags for tasks */
export enum TaskTag {
  Android = "ANDROID",
  Ios = "IOS",
  NodeJs = "NODE_JS",
  Rails = "RAILS",
  React = "REACT",
}

export type UpdateTaskInput = {
  assigneeId?: InputMaybe<Scalars["String"]["input"]>;
  dueDate?: InputMaybe<Scalars["DateTime"]["input"]>;
  id: Scalars["String"]["input"];
  name?: InputMaybe<Scalars["String"]["input"]>;
  pointEstimate?: InputMaybe<PointEstimate>;
  position?: InputMaybe<Scalars["Float"]["input"]>;
  status?: InputMaybe<Status>;
  tags?: InputMaybe<Array<TaskTag>>;
};

export type User = {
  __typename?: "User";
  avatar?: Maybe<Scalars["String"]["output"]>;
  createdAt: Scalars["DateTime"]["output"];
  email: Scalars["String"]["output"];
  fullName: Scalars["String"]["output"];
  id: Scalars["ID"]["output"];
  type: UserType;
  updatedAt: Scalars["DateTime"]["output"];
};

/** Type of the User */
export enum UserType {
  Admin = "ADMIN",
  Candidate = "CANDIDATE",
}

export type TaskFieldsFragment = {
  __typename?: "Task";
  dueDate: any;
  id: string;
  name: string;
  pointEstimate: Types.PointEstimate;
  position: number;
  status: Types.Status;
  tags: Array<Types.TaskTag>;
  assignee?: {
    __typename?: "User";
    avatar?: string | null;
    fullName: string;
    id: string;
  } | null;
};